refactor: #3 광고 섹션 리팩토링
