--- sopt-37th-assignment/sopt-37th-assignment/Global/Component/CustomLayout/AdvertisementCustomLayout.swift
@@ -1,37 +1,0 @@
-//
-//  AdvertisementCustomLayout.swift
-//  sopt-37th-assignment
-//
-//  Created by 이나연 on 11/14/25.
-//
-
-import UIKit
-
-struct AdvertisementCustomLayout {
-    static func createLayout() -> UICollectionViewCompositionalLayout {
-        return UICollectionViewCompositionalLayout { (_, _) in
-            return advdertisementSection
-        }
-    }
-    
-    static var advdertisementSection: NSCollectionLayoutSection {
-        let item = NSCollectionLayoutItem(
-            layoutSize: .init(
-                widthDimension: .fractionalWidth(1),
-                heightDimension: .absolute(114)
-            )
-        )
-        
-        let group = NSCollectionLayoutGroup.horizontal(
-            layoutSize: .init(
-                widthDimension: .fractionalWidth(1),
-                heightDimension: .absolute(114)
-            ),
-            subitems: [item]
-        )
-        
-        let section = NSCollectionLayoutSection(group: group)
-        section.orthogonalScrollingBehavior = .continuous
-        return section
-    }
-}
